Rewrite access log and consent APIs to event APIs
